info box corrected for IPD and removed duplicating components
